# backend/openapi/v1.json
{
  "components": {
    "schemas": {
      "Accepted": {
        "description": "What a call answers when the caller asked not to wait: the task it started.",
        "properties": {
          "job_id": {
            "description": "Follow it with `GET /jobs/{job_id}`, whose `result` holds the report.",
            "type": "string"
          }
        },
        "required": [
          "job_id"
        ],
        "type": "object"
      },
...
      "CapacityForecast": {
        "description": "One destination folder, and the weight of what this plan sends to it.",
        "properties": {
          "fits": {
            "description": "`incoming_bytes` fits in `free_bytes`.",
            "type": "boolean"
          },
          "free_bytes": {
            "format": "int64",
            "type": "integer"
          },
          "incoming_bytes": {
            "description": "Bytes arriving from a *different* filesystem, which is the only traffic\nthat consumes space (see `fits`).",
            "format": "int64",
            "type": "integer"
          },
          "instance_id": {
            "type": "string"
          },
          "instance_name": {
            "type": [
              "string",
              "null"
            ]
          },
          "items": {
            "minimum": 0,
            "type": "integer"
          },
          "path": {
            "type": "string"
          },
          "same_filesystem_bytes": {
            "description": "Bytes arriving from a folder that reports the same free space, and so\nalmost certainly sits on the same filesystem. A move there is a rename\nand costs nothing.",
            "format": "int64",
            "type": "integer"
          }
        },
        "required": [
          "instance_id",
          "path",
          "incoming_bytes",
          "same_filesystem_bytes",
          "free_bytes",
          "items",
          "fits"
        ],
        "type": "object"
      },
...
      "CreateRuleRequest": {
        "additionalProperties": false,
        "description": "A rule as written: what `POST /rules`, `PUT /rules/{id}`, `/rules/validate`\nand `/rules/preview` take.",
        "properties": {
          "conditions": {
            "items": {
              "$ref": "#/components/schemas/Condition"
            },
            "type": "array"
          },
          "description": {
            "type": [
              "string",
              "null"
            ]
          },
          "enabled": {
            "type": "boolean"
          },
          "exclusions": {
            "items": {
              "$ref": "#/components/schemas/Condition"
            },
            "type": "array"
          },
          "instance_ids": {
            "items": {
              "type": "string"
            },
            "type": [
              "array",
              "null"
            ]
          },
          "match_mode": {
            "$ref": "#/components/schemas/MatchMode"
          },
          "media_type": {
            "type": "string"
          },
          "name": {
            "type": "string"
          },
          "priority": {
            "description": "Lower is tried first. Equal priorities go by name, then by id. Left\nout, a new rule goes after every other, and an edited one keeps its\nplace.",
            "format": "int64",
            "type": [
              "integer",
              "null"
            ]
          },
          "target_category": {
            "type": "string"
          }
        },
        "required": [
          "name",
          "media_type",
          "conditions",
          "target_category"
        ],
        "type": "object"
      },
...
      "Deleted": {
        "description": "What a removal answers.",
        "properties": {
          "deleted": {
            "type": "boolean"
          }
        },
        "required": [
          "deleted"
        ],
        "type": "object"
      },
      "ErrorResponse": {
        "description": "The envelope every refusal and failure answers.",
        "properties": {
          "answerable": {
            "description": "Beside `confirm`: whether this caller may send the name back.",
            "type": [
              "boolean",
              "null"
            ]
          },
          "confirm": {
            "description": "The guardrail that asks, for `confirmation_required`. Sending it back\nin `confirm` accepts that question and no other.",
            "type": [
              "string",
              "null"
            ]
          },
          "error": {
            "description": "A stable code: `bad_request`, `unauthorized`, `forbidden`, `not_found`,\n`method_not_allowed`, `conflict`, `confirmation_required`,\n`payload_too_large`, `unsupported_media_type`, `too_many_requests`,\n`busy`, `reauthentication_required`, `passphrase_required`,\n`external_api_error` or `internal_error`. An open list.",
            "type": "string"
          },
          "includes": {
            "description": "Beside `confirm`: the other guardrails the question states, sent back\nin `confirm` with it to accept the question whole.",
            "items": {
              "type": "string"
            },
            "type": [
              "array",
              "null"
            ]
          },
          "message": {
            "description": "A sentence for a person, in the interface language. Never part of the\ncontract.",
            "type": "string"
          },
          "reason": {
            "description": "Why a `conflict` refuses, an open list. A job still running, with a\n`Retry-After` header: `apply_running`, `simulation_running`,\n`previews_full`, `sync_running`, `enrichment_running`, `backup_running`,\n`backup_just_taken`, `restore_staging`, `purge_running`. A state that\nstays until somebody changes it: `name_taken`, `in_use`,\n`already_listed`, `not_declared`, `category_mapped`,\n`proposals_replaced`, `dry_run`, `not_cancellable`, `key_pinned`,\n`only_credential`, `secret_unreadable`.",
            "type": [
              "string",
              "null"
            ]
          }
        },
        "required": [
          "error",
          "message"
        ],
        "type": "object"
      },
...
      "MatchMode": {
        "description": "How the conditions of a rule combine.",
        "enum": [
          "all",
          "any"
        ],
        "type": "string"
      },
...
      "Rule": {
        "description": "A routing rule that maps conditions to a target category.",
        "properties": {
          "conditions": {
            "items": {
              "$ref": "#/components/schemas/Condition"
            },
            "type": "array"
          },
          "created_at": {
            "type": "string"
          },
          "description": {
            "type": [
              "string",
              "null"
            ]
          },
          "enabled": {
            "type": "boolean"
          },
          "exclusions": {
            "description": "Conditions that disqualify the rule even when `conditions` match.",
            "items": {
              "$ref": "#/components/schemas/Condition"
            },
            "type": "array"
          },
          "id": {
            "type": "string"
          },
          "instance_ids": {
            "items": {
              "type": "string"
            },
            "type": [
              "array",
              "null"
            ]
          },
          "match_mode": {
            "$ref": "#/components/schemas/MatchMode"
          },
          "media_type": {
            "type": "string",
            "x-extensible-enum": [
              "movie",
              "series",
              "both"
            ]
          },
          "name": {
            "type": "string"
          },
          "priority": {
            "description": "Lower is tried first. Equal priorities go by name, then by id.",
            "format": "int64",
            "type": "integer"
          },
          "target_category": {
            "type": "string"
          },
          "updated_at": {
            "type": "string"
          }
        },
        "required": [
          "id",
          "name",
          "priority",
          "enabled",
          "media_type",
          "conditions",
          "target_category",
          "created_at",
          "updated_at"
        ],
        "type": "object"
      },
...
      "SimulationRequest": {
        "additionalProperties": false,
        "description": "Request to run a simulation.",
        "properties": {
          "instance_ids": {
            "items": {
              "type": "string"
            },
            "type": [
              "array",
              "null"
            ]
          },
          "max_returned": {
            "description": "Cap the payload size. The counters always reflect the full library.",
            "minimum": 0,
            "type": [
              "integer",
              "null"
            ]
          },
          "media_type": {
            "type": [
              "string",
              "null"
            ]
          },
          "persist": {
            "description": "Store the decisions so they can be applied. `false` is a pure preview.",
            "type": "boolean"
          },
          "persist_unchanged": {
            "description": "Also keep decisions for media already in the right place.",
            "type": "boolean"
          }
        },
        "type": "object"
      },
      "SimulationResult": {
        "description": "Simulation result summary.",
        "properties": {
          "already_correct": {
            "minimum": 0,
            "type": "integer"
          },
          "capacity": {
            "description": "What the plan would put on each destination, and whether it fits.\n\nOnly destinations that receive something appear, and only when the Arr\nreported a free-space figure for them.",
            "items": {
              "$ref": "#/components/schemas/CapacityForecast"
            },
            "type": "array"
          },
          "decisions": {
            "items": {
              "$ref": "#/components/schemas/Decision"
            },
            "type": "array"
          },
          "elapsed_ms": {
            "format": "int64",
            "minimum": 0,
            "type": "integer"
          },
          "excluded_by_rule": {
            "description": "How often a rule matched a title and one of its own exclusions set it\naside, over every title, rules below the winner included.",
            "minimum": 0,
            "type": "integer"
          },
          "moves_required": {
            "minimum": 0,
            "type": "integer"
          },
          "no_category_match": {
            "minimum": 0,
            "type": "integer"
          },
          "overrides_applied": {
            "minimum": 0,
            "type": "integer"
          },
          "returned": {
            "description": "How many decisions are included in `decisions` after truncation.",
            "minimum": 0,
            "type": "integer"
          },
          "simulation_id": {
            "type": "string"
          },
          "skipped_unmapped": {
            "description": "Matched a category with no root folder mapped on that instance.",
            "minimum": 0,
            "type": "integer"
          },
          "total_media": {
            "minimum": 0,
            "type": "integer"
          }
        },
        "required": [
          "simulation_id",
          "total_media",
          "returned",
          "decisions",
          "moves_required",
          "already_correct",
          "no_category_match",
          "overrides_applied",
          "skipped_unmapped",
          "excluded_by_rule",
          "capacity",
          "elapsed_ms"
        ],
        "type": "object"
      },
...
    },
    "securitySchemes": {
      "api_key": {
        "in": "header",
        "name": "X-Api-Key",
        "type": "apiKey"
      },
      "bearer": {
        "scheme": "bearer",
        "type": "http"
      }
    }
  },
  "info": {
...
    "license": {
      "identifier": "GPL-3.0-only",
      "name": "GPL-3.0-only"
    },
    "title": "Routarr API",
    "version": "1"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/rules/{id}": {
      "delete": {
        "operationId": "remove_rule",
        "parameters": [
          {
            "description": "The rule's id.",
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Deleted"
                }
              }
            }
          },
          "default": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "A refusal or a failure, in the envelope every error shares."
          }
        },
        "summary": "Remove a rule",
        "tags": [
          "rules"
        ],
        "x-routarr-scope": "configure"
      },
      "get": {
        "operationId": "get_rule",
        "parameters": [
          {
            "description": "The rule's id.",
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Rule"
                }
              }
            }
          },
          "default": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "A refusal or a failure, in the envelope every error shares."
          }
        },
        "summary": "One rule",
        "tags": [
          "rules"
        ],
        "x-routarr-scope": "read"
      },
      "put": {
        "operationId": "update_rule",
        "parameters": [
          {
            "description": "The rule's id.",
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateRuleRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Rule"
                }
              }
            }
          },
          "default": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "A refusal or a failure, in the envelope every error shares."
          }
        },
        "summary": "Replace a rule",
        "tags": [
          "rules"
        ],
        "x-routarr-scope": "configure"
      }
    },
...
    "/simulate": {
      "post": {
        "description": "With `persist: true` the decisions are stored under the `simulation_id`\nthis answers, replacing the pending ones of an earlier run. Only one\npersisting simulation runs at a time: another answers 409 until it ends.\nA preview, `persist: false`, waits its turn, and answers 409 while four\nare already running or waiting.",
        "operationId": "simulate",
        "parameters": [
          {
            "description": "`respond-async` answers 202 once the task has started, instead of its report.",
            "in": "header",
            "name": "Prefer",
            "required": false,
            "schema": {
              "type": [
                "string",
                "null"
              ]
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SimulationRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SimulationResult"
                }
              }
            }
          },
          "202": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Accepted"
                }
              }
            },
            "description": "The task has started, as `Prefer: respond-async` asked.",
            "headers": {
              "Location": {
                "description": "The task, under `/jobs`.",
                "schema": {
                  "type": "string"
                }
              },
              "Preference-Applied": {
                "description": "`respond-async`.",
                "schema": {
                  "type": "string"
                }
              },
              "Retry-After": {
                "description": "Seconds to wait before the first look at the task.",
                "schema": {
                  "format": "int32",
                  "minimum": 0,
                  "type": "integer"
                }
              }
            }
          },
          "default": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "A refusal or a failure, in the envelope every error shares."
          }
        },
        "summary": "Run the rules over the library",
        "tags": [
          "decisions"
        ],
        "x-routarr-scope": "operate"
      }
    },
...
  },
  "security": [
    {
      "api_key": []
    },
    {
      "bearer": []
    }
  ],
  "servers": [
    {
      "url": "/api/v1"
    }
  ],
...
}